Combine two script files into one file
