+++ lib/frontend/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>tetris_project</title>
-    <script src="../../bazel-bin/lib/frontend/scripts/frame.js"></script>
+    <script src="../../bazel-bin/lib/frontend/scripts/bundle.js"></script>
     <link type="text/css" rel="stylesheet" href="./css/style.css"></link>
 </head>
 <body id="body">
